add breakpoint handling
The one problem is that right now, the messages queue is essentially
synchronous: a message gets sent, and the communication thread waits
for a reply. This means that we can't have a callback that just waits
until a breakpoint is hit after we hit the Resume button: that would
tie up all communications.

Need a better, more async processing.