mostly complete the async debugger GUI work
Every once in a while, we'll get a message that's handled before the
handler is registered. I'm not sure how that happens, since we should
have a happens-before between the "communicate(...)" thread (which
registers the handler before putting the message on the pendingMessages
queue) and the thread that pulls from pendingMessages.

Weirdly enough, it doesn't seem to cause problems. So... eh?