change how method invocations interact with the pc
Before, we would first increment the pc, and then open the frame. That
meant that when the frame closed, it would close right to the next op.
That was nice and simple, but it meant the stack traces were off: each
frame's position was shifted one ahead.

Now, we increment the pc after closing. To do that, we have to seed
the bootstrap frame a bit more, but it's actually a tad _less_ hacky
now. Rather than referring to "no line in no method," it refers to the
only line in the bootstap method.