move EffesOp's generic from return value to param
This makes it more useful for the Effes2 parser. Rather than always
using just Void and not caring about the return value, we can pass in
compiler context like line number and column. This will make it easier
to add debug data, which is going to be increasingly useful.

This is mostly a pretty mechanical change. One thing to maybe note is
that we no longer filter out bridge functions when looking for annotated
methods. This is because in the case of EffesOpsImpl and EffesOps<T>,
the bridge method in EffesOpsImpl is the one that has a corresponding
list of params in EffesOps (namely, the first arg is Object, not
OpBuilder as it is for the reified methods in EffesOpsImpl).